+++ Library/Circle.Library.Business/Handlers/Languages/Commands/DeleteLanguageCommand.cs
@@ -1,7 +1,7 @@
 ﻿using System.Threading;
 using System.Threading.Tasks;
 using Circle.Library.Business.BusinessAspects;
-using Circle.Library.Business.Constants;
+
 using Circle.Core.Aspects.Autofac.Caching;
 using Circle.Core.Aspects.Autofac.Logging;
 using Circle.Core.CrossCuttingConcerns.Logging.Serilog.Loggers;
@@ -12,31 +12,32 @@ using System;
 
 namespace Circle.Library.Business.Handlers.Languages.Commands
 {
-    public class DeleteLanguageCommand : IRequest<IResult>
+    public class DeleteLanguageCommand : IRequest<ResponseMessage<NoContent>>
     {
         public Guid Id { get; set; }
 
-        public class DeleteLanguageCommandHandler : IRequestHandler<DeleteLanguageCommand, IResult>
+        public class DeleteLanguageCommandHandler : IRequestHandler<DeleteLanguageCommand, ResponseMessage<NoContent>>
         {
             private readonly ILanguageRepository _languageRepository;
-            private readonly IMediator _mediator;
 
-            public DeleteLanguageCommandHandler(ILanguageRepository languageRepository, IMediator mediator)
+            public DeleteLanguageCommandHandler(ILanguageRepository languageRepository)
             {
                 _languageRepository = languageRepository;
-                _mediator = mediator;
             }
 
             [CacheRemoveAspect("Get")]
             [LogAspect(typeof(MsSqlLogger))]
             [SecuredOperation(Priority = 1)]
-            public async Task<IResult> Handle(DeleteLanguageCommand request, CancellationToken cancellationToken)
+            public async Task<ResponseMessage<NoContent>> Handle(DeleteLanguageCommand request, CancellationToken cancellationToken)
             {
                 var languageToDelete = _languageRepository.Get(p => p.Id == request.Id);
 
+                if (languageToDelete == null || languageToDelete.Id == Guid.Empty)
+                    return ResponseMessage<NoContent>.NoDataFound("Dil tanımı bulunamadı");
+
                 _languageRepository.Delete(languageToDelete);
                 await _languageRepository.SaveChangesAsync();
-                return new SuccessResult(Messages.Deleted);
+                return ResponseMessage<NoContent>.Success();
             }
         }
     }
